+++ database copy/game.csv
@@ -1,0 +1,1 @@
+id;nama;kategori;tahun_rilis;harga;stok
