+++ database copy/kepemilikan.csv
@@ -1,0 +1,1 @@
+game_id;user_id
