+++ database copy/riwayat.csv
@@ -1,0 +1,1 @@
+game_id;nama;harga;user_id;tahun_beli
